+++ app/types/completedTask.ts
@@ -9,12 +9,12 @@ export interface CompletedTask {
   c_task_id: number;
   user_id: number;
   task_id: number;
+  description: string;
+  payout_value: string;
   created_at: string;
   comment: string | null;
   attachment: string | null;
   payment_status: string;
-  user_name: string;
-  task_title: string;
-  icon_name: string;
-  user_icon: string;
+  task_title?: string;
+  user_name?: string;
 }